+++ source/Core/Autoload/UnifiedNameSpaceClassMap.php
@@ -82,12 +82,6 @@ return [
         'isInterface'      => false,
         'isDeprecated'     => false
     ],
-    'OxidEsales\Eshop\Application\Component\Widget\BetaNote'                     => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Component\Widget\BetaNote::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => false
-    ],
     'OxidEsales\Eshop\Application\Component\Widget\CategoryTree'                 => [
         'editionClassName' => \OxidEsales\EshopCommunity\Application\Component\Widget\CategoryTree::class,
         'isAbstract'       => false,